requirements.txt
jaxlib
jax[cpu]
jupyter
matplotlib
numpy
requests
scipy
seaborn
scikit-learn
h5py
pandas
graphviz
arviz
pyyaml
black
jupyter
nbformat
jupyter-cache
jaxopt
equinox
optax
lineax
torch
lightning
wandb
jaxopt
cola-ml
tqdm
python-fire
jsonargparse[signatures]
gpytorch